+++ packages/network-controller/tests/NetworkController.test.ts
@@ -1579,6 +1579,109 @@ describe('NetworkController', () => {
           });
         });
 
+        describe('if all subscriptions are removed from the messenger before the call to lookupNetwork completes', () => {
+          it('does not throw an error', async () => {
+            const infuraProjectId = 'some-infura-project-id';
+
+            await withController(
+              {
+                state: {
+                  selectedNetworkClientId: infuraNetworkType,
+                },
+                infuraProjectId,
+              },
+              async ({ controller, messenger }) => {
+                const fakeProvider = buildFakeProvider([
+                  // Called during provider initialization
+                  {
+                    request: {
+                      method: 'eth_getBlockByNumber',
+                    },
+                    response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                  },
+                  // Called via `lookupNetwork` directly
+                  {
+                    request: {
+                      method: 'eth_getBlockByNumber',
+                    },
+                    response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                  },
+                ]);
+                const fakeNetworkClient = buildFakeClient(fakeProvider);
+                mockCreateNetworkClient()
+                  .calledWith({
+                    chainId: ChainId[infuraNetworkType],
+                    infuraProjectId,
+                    network: infuraNetworkType,
+                    ticker: NetworksTicker[infuraNetworkType],
+                    type: NetworkClientType.Infura,
+                  })
+                  .mockReturnValue(fakeNetworkClient);
+                await controller.initializeProvider();
+
+                const lookupNetworkPromise = controller.lookupNetwork();
+                messenger.clearSubscriptions();
+                expect(await lookupNetworkPromise).toBeUndefined();
+              },
+            );
+          });
+        });
+
+        describe('if removing the networkDidChange subscription fails for an unknown reason', () => {
+          it('re-throws the error', async () => {
+            const infuraProjectId = 'some-infura-project-id';
+
+            await withController(
+              {
+                state: {
+                  selectedNetworkClientId: infuraNetworkType,
+                },
+                infuraProjectId,
+              },
+              async ({ controller, messenger }) => {
+                const fakeProvider = buildFakeProvider([
+                  // Called during provider initialization
+                  {
+                    request: {
+                      method: 'eth_getBlockByNumber',
+                    },
+                    response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                  },
+                  // Called via `lookupNetwork` directly
+                  {
+                    request: {
+                      method: 'eth_getBlockByNumber',
+                    },
+                    response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                  },
+                ]);
+                const fakeNetworkClient = buildFakeClient(fakeProvider);
+                mockCreateNetworkClient()
+                  .calledWith({
+                    chainId: ChainId[infuraNetworkType],
+                    infuraProjectId,
+                    network: infuraNetworkType,
+                    ticker: NetworksTicker[infuraNetworkType],
+                    type: NetworkClientType.Infura,
+                  })
+                  .mockReturnValue(fakeNetworkClient);
+                await controller.initializeProvider();
+
+                const lookupNetworkPromise = controller.lookupNetwork();
+                const error = new Error('oops');
+                jest
+                  .spyOn(messenger, 'unsubscribe')
+                  .mockImplementation((eventType) => {
+                    if (eventType === 'NetworkController:networkDidChange') {
+                      throw error;
+                    }
+                  });
+                await expect(lookupNetworkPromise).rejects.toThrow(error);
+              },
+            );
+          });
+        });
+
         lookupNetworkTests({
           expectedNetworkClientType: NetworkClientType.Infura,
           initialState: {
@@ -1884,6 +1987,131 @@ describe('NetworkController', () => {
 
               await expect(promiseForNoInfuraIsUnblockedEvents).toBeFulfilled();
               await expect(promiseForInfuraIsBlockedEvents).toBeFulfilled();
+            },
+          );
+        });
+      });
+
+      describe('if all subscriptions are removed from the messenger before the call to lookupNetwork completes', () => {
+        it('does not throw an error', async () => {
+          const infuraProjectId = 'some-infura-project-id';
+
+          await withController(
+            {
+              state: {
+                selectedNetworkClientId: 'AAAA-AAAA-AAAA-AAAA',
+                networkConfigurationsByChainId: {
+                  '0x1337': buildCustomNetworkConfiguration({
+                    chainId: '0x1337',
+                    nativeCurrency: 'TEST',
+                    rpcEndpoints: [
+                      buildCustomRpcEndpoint({
+                        networkClientId: 'AAAA-AAAA-AAAA-AAAA',
+                        url: 'https://test.network',
+                      }),
+                    ],
+                  }),
+                },
+              },
+              infuraProjectId,
+            },
+            async ({ controller, messenger }) => {
+              const fakeProvider = buildFakeProvider([
+                // Called during provider initialization
+                {
+                  request: {
+                    method: 'eth_getBlockByNumber',
+                  },
+                  response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                },
+                // Called via `lookupNetwork` directly
+                {
+                  request: {
+                    method: 'eth_getBlockByNumber',
+                  },
+                  response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                },
+              ]);
+              const fakeNetworkClient = buildFakeClient(fakeProvider);
+              mockCreateNetworkClient()
+                .calledWith({
+                  chainId: '0x1337',
+                  rpcUrl: 'https://test.network',
+                  ticker: 'TEST',
+                  type: NetworkClientType.Custom,
+                })
+                .mockReturnValue(fakeNetworkClient);
+              await controller.initializeProvider();
+
+              const lookupNetworkPromise = controller.lookupNetwork();
+              messenger.clearSubscriptions();
+              await lookupNetworkPromise;
+            },
+          );
+        });
+      });
+
+      describe('if removing the networkDidChange subscription fails for an unknown reason', () => {
+        it('re-throws the error', async () => {
+          const infuraProjectId = 'some-infura-project-id';
+
+          await withController(
+            {
+              state: {
+                selectedNetworkClientId: 'AAAA-AAAA-AAAA-AAAA',
+                networkConfigurationsByChainId: {
+                  '0x1337': buildCustomNetworkConfiguration({
+                    chainId: '0x1337',
+                    nativeCurrency: 'TEST',
+                    rpcEndpoints: [
+                      buildCustomRpcEndpoint({
+                        networkClientId: 'AAAA-AAAA-AAAA-AAAA',
+                        url: 'https://test.network',
+                      }),
+                    ],
+                  }),
+                },
+              },
+              infuraProjectId,
+            },
+            async ({ controller, messenger }) => {
+              const fakeProvider = buildFakeProvider([
+                // Called during provider initialization
+                {
+                  request: {
+                    method: 'eth_getBlockByNumber',
+                  },
+                  response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                },
+                // Called via `lookupNetwork` directly
+                {
+                  request: {
+                    method: 'eth_getBlockByNumber',
+                  },
+                  response: SUCCESSFUL_ETH_GET_BLOCK_BY_NUMBER_RESPONSE,
+                },
+              ]);
+              const fakeNetworkClient = buildFakeClient(fakeProvider);
+              mockCreateNetworkClient()
+                .calledWith({
+                  chainId: '0x1337',
+                  rpcUrl: 'https://test.network',
+                  ticker: 'TEST',
+                  type: NetworkClientType.Custom,
+                })
+                .mockReturnValue(fakeNetworkClient);
+              await controller.initializeProvider();
+
+              const lookupNetworkPromise = controller.lookupNetwork();
+              const error = new Error('oops');
+              jest
+                .spyOn(messenger, 'unsubscribe')
+                .mockImplementation((eventType) => {
+                  if (eventType === 'NetworkController:networkDidChange') {
+                    throw error;
+                  }
+                });
+              await expect(lookupNetworkPromise).rejects.toThrow(error);
             },
           );
         });